Move HmdaSupervisor to parent package
